requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Medical imaging
torchxrayvision>=1.2.0
pydicom>=2.4.0
scikit-image>=0.21.0
Pillow>=10.0.0

# ZCore and embeddings
ftfy>=6.1.0
regex>=2023.0.0
git+https://github.com/openai/CLIP.git
fiftyone>=0.22.0

# Multi-objective optimization
deap>=1.4.0

# Calibration metrics
netcal>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Progress and utilities
tqdm>=4.66.0
pyyaml>=6.0.0
requests>=2.31.0
psutil>=5.9.0
kagglehub>=0.2.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0
